fix(auth): repair failed_login.id autoincrement on SQLite
The original migration declared id as BigInteger; on SQLite, BIGINT
PRIMARY KEY is not a rowid alias and does not autoincrement, so any
INSERT without an explicit id raised NOT NULL. Tests passed because
they use Base.metadata.create_all (which emits INTEGER from the
Mapped[int] model) rather than Alembic.

Fix: correct the original migration to use Integer, and add corrective
migration e9f0a1b2c3d4 (drop + recreate) for existing installs.
The table is a sliding-window throttle log; losing in-flight rows is
acceptable.

+++ src/compendium/migrations/versions/f2a3b4c5d6e7_add_failed_login_table.py
@@ -24,7 +24,7 @@ depends_on: Union[str, Sequence[str], None] = None
 def upgrade() -> None:
     op.create_table(
         "failed_login",
-        sa.Column("id", sa.BigInteger(), nullable=False),
+        sa.Column("id", sa.Integer(), nullable=False),
         sa.Column("scope", sa.Text(), nullable=False),
         sa.Column("identifier", sa.Text(), nullable=False),
         sa.Column("occurred_at", sa.DateTime(timezone=True), nullable=False),
